+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added release-script options to fast-forward and optionally push `master` after the required `origin/dev` push.
 - None
 
 ### Changed